refactor: change file name and folder structure to make compatible with geng
